Enabled Image uploads

+++ database/migrations/2018_07_27_074444_create_applications_table.php
@@ -18,7 +18,7 @@ class CreateApplicationsTable extends Migration
             $table->string('username',50)->unique();
             $table->string('icon',50);
             $table->string('name',50);
-            $table->decimal('owner',10);
+            $table->integer('owner');
             $table->string('description',500);
             $table->timestamps();
         });
